# buscador_inteligente/src/swagger.yaml
openapi: 3.0.0
info:
  title: API de Busca Inteligente NCM
  description: API para consulta de NCM utilizando modelos de IA
  version: 1.0.0
  contact:
    name: Equipe CMEX
servers:
  - url: /api/v1
    description: Servidor de desenvolvimento
  - url: /v1
    description: Servidor para integração Jina UI
...
paths:
  /ncm:
    post:
      tags:
        - NCM
      summary: Consulta a classificação fiscal (NCM) de produtos
      description: Retorna a classificação fiscal (NCM) de um produto com base na descrição fornecida
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NCMRequest"
      responses:
        "200":
          description: Consulta realizada com sucesso
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NCMResponse"
        "400":
          description: Parâmetros inválidos
        "500":
          description: Erro interno do servidor
...
  /api/v1/query:
    post:
      summary: Inicia uma nova consulta à LLM
      tags:
        - Consultas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Query"
      responses:
        "200":
          description: ID da requisição gerado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  requestId:
                    type: string
        "400":
          description: Parâmetros inválidos
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    NCMRequest:
      type: object
      required:
        - consulta
        - estadoOrigem
        - modelo
      properties:
        consulta:
          type: string
          description: Descrição do produto
          example: "Smartphone com tela de 6.5 polegadas, 128GB de armazenamento"
        estadoOrigem:
          type: string
          description: Estado de origem (sigla)
          example: "SP"
        operacao:
          type: string
          description: Tipo de operação
          example: "venda"
        regimeTributario:
          type: string
          description: Regime tributário
          example: "simples"
        tributacao:
          type: string
          description: Tipo de tributação
          example: "normal"
        modelo:
          type: string
          description: Modelo de IA a ser utilizado
          example: "gemini-2.0-flash"
        useDeepResearch:
          type: boolean
          description: Ativa o modo de análise profunda
          default: false
    NCMResponse:
      type: object
      properties:
        ncm:
          type: string
          description: Código NCM
          example: "8517.12.31"
        descricao:
          type: string
          description: Descrição do NCM
          example: "Telefones móveis (smartphones)"
        atributos:
          type: array
          items:
            type: string
          description: Atributos do produto
          example: ["Comunicação sem fio", "Tela touchscreen"]
        atributos_tipi:
          type: array
          items:
            type: string
          description: Atributos específicos da TIPI
          example: ["Ex 01 - Smartphones"]
        valores_de_impostos:
          type: object
          properties:
            ipi:
              type: string
              description: Valor do IPI
              example: "15%"
            icms:
              type: object
              additionalProperties:
                type: string
              description: Valores de ICMS por estado
              example: { "SP": "18%", "RJ": "20%" }
            pis:
              type: string
              description: Valor do PIS
              example: "1.65%"
            cofins:
              type: string
              description: Valor do COFINS
              example: "7.6%"
...
    Query:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          description: Consulta ao modelo de IA
          example: "Qual a classificação fiscal para um smartphone?"
...